Extract the request identity boundary
Agent-Assisted-By: Codex (GPT-5)
